implement solution for checking kids with greatest number of candies
Added a solution for the LeetCode problem 1431 and committed it to GitHub as part of my coding practice. The function determines whether each child can have the greatest number of candies if given a certain number of extra candies. This was a good exercise to practice array manipulation and condition checking in JavaScript.